+++ service-worker/alarms.ts
@@ -3,3 +3,5 @@ export const Alarms = {
   SearchForOpenQueueFiles: 'searchForOpenQueueFiles',
   DeleteOpenQueueFiles: 'deleteOpenQueueFiles',
 } as const
+
+export type AlarmsT = (typeof Alarms)[keyof typeof Alarms]