+++ service-worker/index.ts
@@ -1,15 +1,19 @@
-import {Alarms} from './alarms'
+import {Alarms, AlarmsT} from './alarms'
 import {backupOpenWindows} from './backup'
 import {deleteOpenQueueFiles} from './delete'
 import {loadBackupSubDirectory, loadDirectoryHandle} from './load'
 import {searchForOpenQueueFiles} from './open'
 import {store} from './store'
 import {backgroundLog} from './utils/log'
+import {Messages, MessagesT} from './utils/message'
+
+let exhaustiveCheck: never
 
 chrome.action.onClicked.addListener(() => chrome.runtime.openOptionsPage())
 
-chrome.alarms.onAlarm.addListener(alarm => {
-  switch (alarm.name) {
+chrome.alarms.onAlarm.addListener((alarm: chrome.alarms.Alarm) => {
+  const alarmName = alarm.name as AlarmsT
+  switch (alarmName) {
     case Alarms.BackupOpenWindows:
       backupOpenWindows()
       break
@@ -20,25 +24,29 @@ chrome.alarms.onAlarm.addListener(alarm => {
       deleteOpenQueueFiles()
       break
     default:
-      backgroundLog('Unexpected alarm: ', alarm.name)
+      exhaustiveCheck = alarmName
+      backgroundLog('Unexpected alarm: ', exhaustiveCheck)
   }
 })
 
-chrome.runtime.onMessage.addListener(message => {
+chrome.runtime.onMessage.addListener((message: MessagesT) => {
   switch (message) {
-    case 'Changed Directory Handle':
+    case Messages.ChangedDirectoryHandle:
       init()
       break
-    case 'Changed Backup Directory':
+    case Messages.ChangedBackupDirectory:
       loadBackupSubDirectory()
       break
-    case 'Manually Run Backup':
-      backupOpenWindows()
+    case Messages.ManuallyRunBackup:
+      backupOpenWindows().then(() => chrome.runtime.sendMessage(Messages.BackupCompleted))
       break
-    case 'Request Permission':
+    case Messages.RequestPermission:
+      break
+    case Messages.BackupCompleted:
       break
     default:
-      backgroundLog('Unexpected message: ', message)
+      exhaustiveCheck = message
+      backgroundLog('Unexpected message: ', exhaustiveCheck)
   }
 })