+++ src/components/backup.tsx
@@ -1,13 +1,22 @@
 import {Dispatch, FormEvent, SetStateAction, useState} from 'react'
+import {Messages} from '../utils/message'
 import {storeBackupDirectory} from '../utils/store-backup-directory'
 
 type Props = {
   backupDirectory: string
   setBackupDirectory: Dispatch<SetStateAction<string>>
   clearBackupDirectory: () => void
+  backupLoading: boolean
+  setBackupLoading: (backupLoading: boolean) => void
 }
 
-export function Backup({backupDirectory, setBackupDirectory, clearBackupDirectory}: Props) {
+export function Backup({
+  backupDirectory,
+  setBackupDirectory,
+  clearBackupDirectory,
+  backupLoading,
+  setBackupLoading,
+}: Props) {
   const [backupInputText, setBackupInputText] = useState('')
   const [isSettingUpBackup, setIsSettingUpBackup] = useState(false)
 
@@ -35,10 +44,14 @@ export function Backup({backupDirectory, setBackupDirectory, clearBackupDirector
       </div>
       <div className="row">
         <button
-          onClick={() => chrome.runtime.sendMessage('Manually Run Backup')}
+          onClick={() => {
+            setBackupLoading(true)
+            chrome.runtime.sendMessage(Messages.ManuallyRunBackup)
+          }}
           className="wide-500"
+          disabled={backupLoading}
         >
-          Manually Run Backup
+          {backupLoading ? 'Backup in Progress...' : 'Manually Run Backup'}
         </button>
       </div>
       <div className="row">